private event request

+++ ewm/src/main/java/ru/practicum/ewm/dto/eventRequest/EventRequestStatusUpdateRequest.java
@@ -11,6 +11,6 @@ import java.util.List;
 @Setter
 @AllArgsConstructor
 public class EventRequestStatusUpdateRequest {
-    private List<Integer> requestIds;
+    private List<Long> requestIds;
     private EventRequestStatus status;
 }
